Tidy GetDockerOS() function
